\name{QueueSimulator-package}
\alias{QueueSimulator-package}
\alias{QueueSimulator}
\docType{package}
\title{
\packageTitle{QueueSimulator}
}
\description{
\packageDescription{QueueSimulator}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QueueSimulator}
\packageIndices{QueueSimulator}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{QueueSimulator}

Maintainer: \packageMaintainer{QueueSimulator}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

